Moved files into a more idiomatic Ruby directory hierarchy.
