+++ lib/xap/parser.rb
@@ -1,0 +1,6 @@
+module Xap
+	module Parser
+	end
+end
+
+require_relative 'parser/parse_xap'
