+++ lib/xap/schema.rb
@@ -1,0 +1,7 @@
+module Xap
+	module Schema
+	end
+end
+
+require_relative 'schema/xap_bsc'
+require_relative 'schema/xap_bsc_dev'
