+++ lib/xap_ruby.rb
@@ -1,5 +1,6 @@
-require "xap_ruby/version"
+require 'xap_ruby/version'
+require 'xap'
 
+# Note: all functions live in the Xap module.
 module XapRuby
-  # Your code goes here...
 end
