+++ test/xap_receive.rb
@@ -2,12 +2,8 @@
 # Displays all xAP messages received.
 # (C)2012 Mike Bourgeous
 
-require 'eventmachine'
-
-path = File.expand_path(File.dirname(__FILE__))
-require File.join(path, '..', 'xap.rb')
-require File.join(path, '..', 'schema', 'xap_bsc.rb')
-require File.join(path, '..', 'schema', 'xap_bsc_dev.rb')
+require 'bundler/setup'
+require 'xap_ruby'
 
 if __FILE__ == $0
 	EM::run {
@@ -22,10 +18,10 @@ if __FILE__ == $0
 			Xap.log "\e[1mIncoming message: \e[34m#{msg.class.name}\e[0m\n\t#{msg.inspect.lines.to_a.join("\t")}\n"
 		}
 
-		Xap.add_receiver XapAddress.parse('*.*.>'), cb
+		Xap.add_receiver Xap::XapAddress.parse('*.*.>'), cb
 
 		trap "EXIT" do
-			Xap.remove_receiver XapAddress.parse('*.*.>'), cb
+			Xap.remove_receiver Xap::XapAddress.parse('*.*.>'), cb
 			Xap.stop_xap
 		end	
 	}
